requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
discord==1.0.1
discord.py==1.6.0
docopt==0.6.2
emoji==1.2.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.3.4
monotonic==1.6
multidict==6.0.2
numpy==1.22.3
pandas==1.2.2
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
python-valve==0.2.1
pytz==2022.1
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
valve==0.0.0
yarl==1.7.2
